Add getOrProvisionSKR
