+++ tests/fast-integration/skr-test/test.js
@@ -1,23 +1,14 @@
 const {
   gatherOptions,
-  withSuffix,
-  withInstanceID,
-  provisionSKRInstance,
   director,
   commerceMockTest,
   oidcE2ETest,
 } = require('./index');
 const {
-  getEnvOrThrow,
-  genRandom,
 } = require('../utils');
 const {unregisterKymaFromCompass} = require('../compass');
 const {deprovisionSKRInstance} = require('./provision/deprovision-skr');
-const {
-  getSKRConfig,
-  prepareCompassResources,
-  initK8sConfig,
-} = require('./helpers');
+const {getOrProvisionSKR} = require('./provision/provision-skr');
 
 const skipProvisioning = process.env.SKIP_PROVISIONING === 'true';
 const provisioningTimeout = 1000 * 60 * 30; // 30m
@@ -39,28 +30,7 @@ describe('SKR test', function() {
   };
 
   before('Ensure SKR is provisioned', async function() {
-    this.timeout(provisioningTimeout);
-    if (skipProvisioning) {
-      console.log('Gather information from externally provisioned SKR and prepare the compass resources');
-      const instanceID = getEnvOrThrow('INSTANCE_ID');
-      let suffix = process.env.TEST_SUFFIX;
-      if (suffix === undefined) {
-        suffix = genRandom(4);
-      }
-      options = gatherOptions(
-          withInstanceID(instanceID),
-          withSuffix(suffix),
-      );
-      shoot = await getSKRConfig(instanceID);
-    } else {
-      console.log('Provisioning new SKR instance...');
-      shoot = await provisionSKRInstance(options, provisioningTimeout);
-    }
-    console.log('Preparing compass resources on the SKR instance...');
-    await prepareCompassResources(shoot, options);
-
-    console.log('Initiating K8s config...');
-    await initK8sConfig(shoot);
+    await getOrProvisionSKR(options, shoot, skipProvisioning, provisioningTimeout);
   });
 
   // Run the OIDC tests
